refactor(chalice): use CH everywhere to get error's details refactor(chalice): upgraded dependencies
