+++ api/chalicelib/core/issues/__init__.py
@@ -1,11 +1,5 @@
 import logging
 
-from decouple import config
-
 logger = logging.getLogger(__name__)
 
-if config("EXP_EVENTS", cast=bool, default=False):
-    logger.info(">>> Using experimental issues")
-    from . import issues_ch as issues
-else:
-    from . import issues_pg as issues
+from . import issues_ch as issues
