+++ ee/api/requirements.txt
@@ -1,15 +1,15 @@
 urllib3==2.5.0
-requests==2.32.4
-boto3==1.39.2
+requests==2.32.5
+boto3==1.40.27
 pyjwt==2.10.1
 psycopg2-binary==2.9.10
-psycopg[pool,binary]==3.2.9
+psycopg[pool,binary]==3.2.10
 clickhouse-connect==0.8.18
-elasticsearch==9.0.2
-jira==3.8.0
-cachetools==6.1.0
+elasticsearch==9.1.0
+jira==3.10.5
+cachetools==6.2.0
 
-fastapi==0.115.14
+fastapi==0.116.1
 uvicorn[standard]==0.35.0
 gunicorn==23.0.0
 python-decouple==3.8
@@ -24,6 +24,6 @@ xmlsec==1.3.14 --no-binary=xmlsec
 
 python-multipart==0.0.20
 
-redis==6.2.0
+redis==6.4.0
 #confluent-kafka==2.1.0
 azure-storage-blob==12.25.1
